refactor(config): move focused helpers into config submodules
Extract automation, blocklist, diagnostics, migration, shortcut, and WakaTime config helpers from the facade.

Preserve the public config schema and facade exports while making each concern easier to navigate and test.
