+++ src/config.rs
@@ -1,19 +1,51 @@
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 use std::ffi::OsString;
 use std::fs;
 use std::io;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
-use crate::blocker::{domain_rule_matches_host, normalize_domain_rule};
 use chrono::{Local, NaiveDate};
 use serde::{Deserialize, Deserializer, Serialize};
 
+mod automation;
+mod blocklists;
+mod diagnostics;
+mod migration;
 mod paths;
+mod shortcuts;
+mod wakatime;
+
+pub use automation::validate_automation_trigger_rules;
+use automation::{
+    normalize_automation_triggers, normalize_trigger_days, normalize_weekday_profile_rules,
+    normalize_weekday_token,
+};
+pub use blocklists::{
+    BlocklistCategoryConfig, BlocklistProfileConfig, effective_blocked_sites_for_profile,
+};
+use blocklists::{
+    default_blocklist_profile_name, make_unique_profile_name, normalize_blocklist_profiles,
+    normalize_selected_blocklist_profile,
+};
+pub use diagnostics::{run_config_doctor, run_config_migration_assistant};
+#[cfg(test)]
+use diagnostics::{run_config_doctor_with_path, run_config_migration_assistant_with_path};
+#[cfg(test)]
+use migration::migrate_config_toml_legacy_to_v1;
+use migration::{
+    collect_legacy_profile_rename_advice, config_health_error, config_health_warning,
+    detect_legacy_config_deprecation_warnings, migrate_config_toml_to_current,
+    migrate_config_toml_to_current_detailed, sort_config_health_findings, summarize_config_health,
+};
 #[cfg(all(test, not(target_os = "windows")))]
 use paths::env_path_from_value;
 use paths::{app_dir_with_env, stats_app_dir_with_env};
 #[cfg(test)]
 use paths::{config_dir_from_env, stats_state_dir_from_env};
+pub use shortcuts::ShortcutConfig;
+#[allow(unused_imports)]
+pub use wakatime::WakatimeTaskMappingConfig;
+pub use wakatime::{WakatimeMetadataConfig, WakatimeRuntimeConfig};
 
 const CURRENT_CONFIG_SCHEMA_VERSION: u32 = 2;
 const LEGACY_CONFIG_SCHEMA_VERSION: u32 = 0;
@@ -25,13 +57,6 @@ const CALENDAR_SYNC_REFRESH_MIN_SECS: u64 = 5 * 60;
 const CALENDAR_SYNC_REFRESH_MAX_SECS: u64 = 24 * 60 * 60;
 const CALENDAR_SYNC_LOOKAHEAD_MIN_DAYS: u16 = 1;
 const CALENDAR_SYNC_LOOKAHEAD_MAX_DAYS: u16 = 90;
-const WAKATIME_QUEUE_CAPACITY_MIN: usize = 1;
-const WAKATIME_QUEUE_CAPACITY_MAX: usize = 4096;
-const WAKATIME_RETRY_DELAY_MIN_SECS: u64 = 1;
-const WAKATIME_RETRY_DELAY_MAX_SECS: u64 = 60 * 60;
-const WAKATIME_RETRY_BACKOFF_MIN_SECS: u64 = 1;
-const WAKATIME_RETRY_BACKOFF_MAX_SECS: u64 = 300;
-const WAKATIME_RETRY_BACKOFF_MAX_ENTRIES: usize = 8;
 
 /// Persistent application configuration stored as TOML.
 ///
@@ -325,472 +350,6 @@ impl CommandBlockingBackendConfig {
             diagnostics_command: self.diagnostics_command.trim().to_string(),
         }
     }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ShortcutConfig {
-    #[serde(default = "default_shortcut_quit")]
-    pub quit: String,
-    #[serde(default = "default_shortcut_timer_toggle_pause")]
-    pub timer_toggle_pause: String,
-    #[serde(default = "default_shortcut_timer_stop_reset")]
-    pub timer_stop_reset: String,
-    #[serde(default = "default_shortcut_timer_next_phase")]
-    pub timer_next_phase: String,
-    #[serde(default = "default_shortcut_open_site_manager")]
-    pub open_site_manager: String,
-    #[serde(default = "default_shortcut_open_profile_manager")]
-    pub open_profile_manager: String,
-    #[serde(default = "default_shortcut_open_session_planner")]
-    pub open_session_planner: String,
-    #[serde(default = "default_shortcut_open_stats_history")]
-    pub open_stats_history: String,
-    #[serde(default = "default_shortcut_open_setup_diagnostics")]
-    pub open_setup_diagnostics: String,
-    #[serde(default = "default_shortcut_timer_edit_note")]
-    pub timer_edit_note: String,
-    #[serde(default = "default_shortcut_break_glass_override")]
-    pub break_glass_override: String,
-    #[serde(default = "default_shortcut_delay_schedule_start")]
-    pub delay_schedule_start: String,
-    #[serde(default = "default_shortcut_back_site_manager")]
-    pub back_site_manager: String,
-    #[serde(default = "default_shortcut_toggle_site_list_mode")]
-    pub toggle_site_list_mode: String,
-    #[serde(default = "default_shortcut_site_add")]
-    pub site_add: String,
-    #[serde(default = "default_shortcut_site_edit")]
-    pub site_edit: String,
-    #[serde(default = "default_shortcut_site_delete")]
-    pub site_delete: String,
-    #[serde(default = "default_shortcut_select_previous_blocklist_profile")]
-    pub select_previous_blocklist_profile: String,
-    #[serde(default = "default_shortcut_select_next_blocklist_profile")]
-    pub select_next_blocklist_profile: String,
-    #[serde(default = "default_shortcut_create_blocklist_profile")]
-    pub create_blocklist_profile: String,
-    #[serde(default = "default_shortcut_rename_blocklist_profile")]
-    pub rename_blocklist_profile: String,
-    #[serde(default = "default_shortcut_delete_blocklist_profile")]
-    pub delete_blocklist_profile: String,
-    #[serde(default = "default_shortcut_back_session_planner")]
-    pub back_session_planner: String,
-    #[serde(default = "default_shortcut_planner_add")]
-    pub planner_add: String,
-    #[serde(default = "default_shortcut_planner_rename")]
-    pub planner_rename: String,
-    #[serde(default = "default_shortcut_planner_favorite")]
-    pub planner_favorite: String,
-    #[serde(default = "default_shortcut_planner_archive")]
-    pub planner_archive: String,
-    #[serde(default = "default_shortcut_planner_delete")]
-    pub planner_delete: String,
-    #[serde(default = "default_shortcut_planner_select_recent")]
-    pub planner_select_recent: String,
-    #[serde(default = "default_shortcut_back_profile_manager")]
-    pub back_profile_manager: String,
-    #[serde(default = "default_shortcut_profile_edit")]
-    pub profile_edit: String,
-    #[serde(default = "default_shortcut_back_stats_history")]
-    pub back_stats_history: String,
-    #[serde(default = "default_shortcut_export_stats_history")]
-    pub export_stats_history: String,
-    #[serde(default = "default_shortcut_history_dashboard_select_previous")]
-    pub history_dashboard_select_previous: String,
-    #[serde(default = "default_shortcut_history_dashboard_select_next")]
-    pub history_dashboard_select_next: String,
-    #[serde(default = "default_shortcut_history_dashboard_toggle_pin")]
-    pub history_dashboard_toggle_pin: String,
-    #[serde(default = "default_shortcut_history_dashboard_move_left")]
-    pub history_dashboard_move_left: String,
-    #[serde(default = "default_shortcut_history_dashboard_move_right")]
-    pub history_dashboard_move_right: String,
-    #[serde(default = "default_shortcut_back_setup_diagnostics")]
-    pub back_setup_diagnostics: String,
-    #[serde(default = "default_shortcut_refresh_setup_diagnostics")]
-    pub refresh_setup_diagnostics: String,
-    #[serde(default = "default_shortcut_navigate_up")]
-    pub navigate_up: String,
-    #[serde(default = "default_shortcut_navigate_down")]
-    pub navigate_down: String,
-    #[serde(default = "default_shortcut_navigate_left")]
-    pub navigate_left: String,
-    #[serde(default = "default_shortcut_navigate_right")]
-    pub navigate_right: String,
-    #[serde(default = "default_shortcut_confirm")]
-    pub confirm: String,
-    #[serde(default = "default_shortcut_cancel")]
-    pub cancel: String,
-    #[serde(default = "default_shortcut_delete")]
-    pub delete: String,
-    #[serde(default = "default_shortcut_backspace")]
-    pub backspace: String,
-}
-
-impl ShortcutConfig {
-    pub fn normalized(&self) -> Self {
-        Self {
-            quit: normalize_shortcut_token(&self.quit, &default_shortcut_quit()),
-            timer_toggle_pause: normalize_shortcut_token(
-                &self.timer_toggle_pause,
-                &default_shortcut_timer_toggle_pause(),
-            ),
-            timer_stop_reset: normalize_shortcut_token(
-                &self.timer_stop_reset,
-                &default_shortcut_timer_stop_reset(),
-            ),
-            timer_next_phase: normalize_shortcut_token(
-                &self.timer_next_phase,
-                &default_shortcut_timer_next_phase(),
-            ),
-            open_site_manager: normalize_shortcut_token(
-                &self.open_site_manager,
-                &default_shortcut_open_site_manager(),
-            ),
-            open_profile_manager: normalize_shortcut_token(
-                &self.open_profile_manager,
-                &default_shortcut_open_profile_manager(),
-            ),
-            open_session_planner: normalize_shortcut_token(
-                &self.open_session_planner,
-                &default_shortcut_open_session_planner(),
-            ),
-            open_stats_history: normalize_shortcut_token(
-                &self.open_stats_history,
-                &default_shortcut_open_stats_history(),
-            ),
-            open_setup_diagnostics: normalize_shortcut_token(
-                &self.open_setup_diagnostics,
-                &default_shortcut_open_setup_diagnostics(),
-            ),
-            timer_edit_note: normalize_shortcut_token(
-                &self.timer_edit_note,
-                &default_shortcut_timer_edit_note(),
-            ),
-            break_glass_override: normalize_shortcut_token(
-                &self.break_glass_override,
-                &default_shortcut_break_glass_override(),
-            ),
-            delay_schedule_start: normalize_shortcut_token(
-                &self.delay_schedule_start,
-                &default_shortcut_delay_schedule_start(),
-            ),
-            back_site_manager: normalize_shortcut_token(
-                &self.back_site_manager,
-                &default_shortcut_back_site_manager(),
-            ),
-            toggle_site_list_mode: normalize_shortcut_token(
-                &self.toggle_site_list_mode,
-                &default_shortcut_toggle_site_list_mode(),
-            ),
-            site_add: normalize_shortcut_token(&self.site_add, &default_shortcut_site_add()),
-            site_edit: normalize_shortcut_token(&self.site_edit, &default_shortcut_site_edit()),
-            site_delete: normalize_shortcut_token(
-                &self.site_delete,
-                &default_shortcut_site_delete(),
-            ),
-            select_previous_blocklist_profile: normalize_shortcut_token(
-                &self.select_previous_blocklist_profile,
-                &default_shortcut_select_previous_blocklist_profile(),
-            ),
-            select_next_blocklist_profile: normalize_shortcut_token(
-                &self.select_next_blocklist_profile,
-                &default_shortcut_select_next_blocklist_profile(),
-            ),
-            create_blocklist_profile: normalize_shortcut_token(
-                &self.create_blocklist_profile,
-                &default_shortcut_create_blocklist_profile(),
-            ),
-            rename_blocklist_profile: normalize_shortcut_token(
-                &self.rename_blocklist_profile,
-                &default_shortcut_rename_blocklist_profile(),
-            ),
-            delete_blocklist_profile: normalize_shortcut_token(
-                &self.delete_blocklist_profile,
-                &default_shortcut_delete_blocklist_profile(),
-            ),
-            back_session_planner: normalize_shortcut_token(
-                &self.back_session_planner,
-                &default_shortcut_back_session_planner(),
-            ),
-            planner_add: normalize_shortcut_token(
-                &self.planner_add,
-                &default_shortcut_planner_add(),
-            ),
-            planner_rename: normalize_shortcut_token(
-                &self.planner_rename,
-                &default_shortcut_planner_rename(),
-            ),
-            planner_favorite: normalize_shortcut_token(
-                &self.planner_favorite,
-                &default_shortcut_planner_favorite(),
-            ),
-            planner_archive: normalize_shortcut_token(
-                &self.planner_archive,
-                &default_shortcut_planner_archive(),
-            ),
-            planner_delete: normalize_shortcut_token(
-                &self.planner_delete,
-                &default_shortcut_planner_delete(),
-            ),
-            planner_select_recent: normalize_shortcut_token(
-                &self.planner_select_recent,
-                &default_shortcut_planner_select_recent(),
-            ),
-            back_profile_manager: normalize_shortcut_token(
-                &self.back_profile_manager,
-                &default_shortcut_back_profile_manager(),
-            ),
-            profile_edit: normalize_shortcut_token(
-                &self.profile_edit,
-                &default_shortcut_profile_edit(),
-            ),
-            back_stats_history: normalize_shortcut_token(
-                &self.back_stats_history,
-                &default_shortcut_back_stats_history(),
-            ),
-            export_stats_history: normalize_shortcut_token(
-                &self.export_stats_history,
-                &default_shortcut_export_stats_history(),
-            ),
-            history_dashboard_select_previous: normalize_shortcut_token(
-                &self.history_dashboard_select_previous,
-                &default_shortcut_history_dashboard_select_previous(),
-            ),
-            history_dashboard_select_next: normalize_shortcut_token(
-                &self.history_dashboard_select_next,
-                &default_shortcut_history_dashboard_select_next(),
-            ),
-            history_dashboard_toggle_pin: normalize_shortcut_token(
-                &self.history_dashboard_toggle_pin,
-                &default_shortcut_history_dashboard_toggle_pin(),
-            ),
-            history_dashboard_move_left: normalize_shortcut_token(
-                &self.history_dashboard_move_left,
-                &default_shortcut_history_dashboard_move_left(),
-            ),
-            history_dashboard_move_right: normalize_shortcut_token(
-                &self.history_dashboard_move_right,
-                &default_shortcut_history_dashboard_move_right(),
-            ),
-            back_setup_diagnostics: normalize_shortcut_token(
-                &self.back_setup_diagnostics,
-                &default_shortcut_back_setup_diagnostics(),
-            ),
-            refresh_setup_diagnostics: normalize_shortcut_token(
-                &self.refresh_setup_diagnostics,
-                &default_shortcut_refresh_setup_diagnostics(),
-            ),
-            navigate_up: normalize_navigation_shortcut_token(
-                &self.navigate_up,
-                &default_shortcut_navigate_up(),
-            ),
-            navigate_down: normalize_navigation_shortcut_token(
-                &self.navigate_down,
-                &default_shortcut_navigate_down(),
-            ),
-            navigate_left: normalize_navigation_shortcut_token(
-                &self.navigate_left,
-                &default_shortcut_navigate_left(),
-            ),
-            navigate_right: normalize_navigation_shortcut_token(
-                &self.navigate_right,
-                &default_shortcut_navigate_right(),
-            ),
-            confirm: normalize_navigation_shortcut_token(
-                &self.confirm,
-                &default_shortcut_confirm(),
-            ),
-            cancel: normalize_navigation_shortcut_token(&self.cancel, &default_shortcut_cancel()),
-            delete: normalize_navigation_shortcut_token(&self.delete, &default_shortcut_delete()),
-            backspace: normalize_navigation_shortcut_token(
-                &self.backspace,
-                &default_shortcut_backspace(),
-            ),
-        }
-    }
-}
-
-impl Default for ShortcutConfig {
-    fn default() -> Self {
-        Self {
-            quit: default_shortcut_quit(),
-            timer_toggle_pause: default_shortcut_timer_toggle_pause(),
-            timer_stop_reset: default_shortcut_timer_stop_reset(),
-            timer_next_phase: default_shortcut_timer_next_phase(),
-            open_site_manager: default_shortcut_open_site_manager(),
-            open_profile_manager: default_shortcut_open_profile_manager(),
-            open_session_planner: default_shortcut_open_session_planner(),
-            open_stats_history: default_shortcut_open_stats_history(),
-            open_setup_diagnostics: default_shortcut_open_setup_diagnostics(),
-            timer_edit_note: default_shortcut_timer_edit_note(),
-            break_glass_override: default_shortcut_break_glass_override(),
-            delay_schedule_start: default_shortcut_delay_schedule_start(),
-            back_site_manager: default_shortcut_back_site_manager(),
-            toggle_site_list_mode: default_shortcut_toggle_site_list_mode(),
-            site_add: default_shortcut_site_add(),
-            site_edit: default_shortcut_site_edit(),
-            site_delete: default_shortcut_site_delete(),
-            select_previous_blocklist_profile: default_shortcut_select_previous_blocklist_profile(),
-            select_next_blocklist_profile: default_shortcut_select_next_blocklist_profile(),
-            create_blocklist_profile: default_shortcut_create_blocklist_profile(),
-            rename_blocklist_profile: default_shortcut_rename_blocklist_profile(),
-            delete_blocklist_profile: default_shortcut_delete_blocklist_profile(),
-            back_session_planner: default_shortcut_back_session_planner(),
-            planner_add: default_shortcut_planner_add(),
-            planner_rename: default_shortcut_planner_rename(),
-            planner_favorite: default_shortcut_planner_favorite(),
-            planner_archive: default_shortcut_planner_archive(),
-            planner_delete: default_shortcut_planner_delete(),
-            planner_select_recent: default_shortcut_planner_select_recent(),
-            back_profile_manager: default_shortcut_back_profile_manager(),
-            profile_edit: default_shortcut_profile_edit(),
-            back_stats_history: default_shortcut_back_stats_history(),
-            export_stats_history: default_shortcut_export_stats_history(),
-            history_dashboard_select_previous: default_shortcut_history_dashboard_select_previous(),
-            history_dashboard_select_next: default_shortcut_history_dashboard_select_next(),
-            history_dashboard_toggle_pin: default_shortcut_history_dashboard_toggle_pin(),
-            history_dashboard_move_left: default_shortcut_history_dashboard_move_left(),
-            history_dashboard_move_right: default_shortcut_history_dashboard_move_right(),
-            back_setup_diagnostics: default_shortcut_back_setup_diagnostics(),
-            refresh_setup_diagnostics: default_shortcut_refresh_setup_diagnostics(),
-            navigate_up: default_shortcut_navigate_up(),
-            navigate_down: default_shortcut_navigate_down(),
-            navigate_left: default_shortcut_navigate_left(),
-            navigate_right: default_shortcut_navigate_right(),
-            confirm: default_shortcut_confirm(),
-            cancel: default_shortcut_cancel(),
-            delete: default_shortcut_delete(),
-            backspace: default_shortcut_backspace(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct BlocklistProfileConfig {
-    #[serde(default = "default_blocklist_profile_name")]
-    pub name: String,
-    /// Deprecated flat mirror of categorized blocklist rules.
-    ///
-    /// Canonical rules live in `categories[*].sites`; this field is maintained
-    /// for load-time compatibility and helper surfaces still reading flat lists.
-    #[serde(default)]
-    pub sites: Vec<String>,
-    /// Deprecated flat mirror of categorized allowlist rules.
-    ///
-    /// Canonical rules live in `categories[*].allowlist_sites`.
-    /// Sites that are explicitly excluded from blocking.
-    ///
-    /// Effective focus blocking is computed as `sites - allowlist_sites`.
-    #[serde(default)]
-    pub allowlist_sites: Vec<String>,
-    /// Category-organized block/allow rules.
-    #[serde(default)]
-    pub categories: Vec<BlocklistCategoryConfig>,
-    /// Name of the selected category inside this profile.
-    #[serde(default = "default_blocklist_category_name")]
-    pub selected_category: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct BlocklistCategoryConfig {
-    #[serde(default = "default_blocklist_category_name")]
-    pub name: String,
-    #[serde(default)]
-    pub sites: Vec<String>,
-    #[serde(default)]
-    pub allowlist_sites: Vec<String>,
-}
-
-impl Default for BlocklistProfileConfig {
-    fn default() -> Self {
-        Self {
-            name: default_blocklist_profile_name(),
-            sites: Vec::new(),
-            allowlist_sites: Vec::new(),
-            categories: Vec::new(),
-            selected_category: default_blocklist_category_name(),
-        }
-    }
-}
-
-impl Default for BlocklistCategoryConfig {
-    fn default() -> Self {
-        Self {
-            name: default_blocklist_category_name(),
-            sites: Vec::new(),
-            allowlist_sites: Vec::new(),
-        }
-    }
-}
-
-pub fn effective_blocked_sites_for_profile(profile: &BlocklistProfileConfig) -> Vec<String> {
-    let allowlist_rules: Vec<String> = all_allowlist_rules_for_profile(profile)
-        .iter()
-        .filter_map(|rule| normalize_domain_rule(rule).ok())
-        .collect();
-
-    let mut seen = HashSet::new();
-    all_blocklist_rules_for_profile(profile)
-        .into_iter()
-        .filter_map(|site| normalize_domain_rule(&site).ok())
-        .filter(|site| !block_rule_excluded_by_allowlist(site, &allowlist_rules))
-        .filter(|site| seen.insert(site.to_ascii_lowercase()))
-        .collect()
-}
-
-fn block_rule_excluded_by_allowlist(block_rule: &str, allowlist_rules: &[String]) -> bool {
-    if let Some(block_suffix) = block_rule.strip_prefix("*.") {
-        let block_suffix = block_suffix.to_ascii_lowercase();
-        return allowlist_rules.iter().any(|allow_rule| {
-            if allow_rule.eq_ignore_ascii_case(block_rule) {
-                return true;
-            }
-            let Some(allow_suffix) = allow_rule.strip_prefix("*.") else {
-                return false;
-            };
-            let allow_suffix = allow_suffix.to_ascii_lowercase();
-            block_suffix == allow_suffix || block_suffix.ends_with(&format!(".{allow_suffix}"))
-        });
-    }
-    allowlist_rules
-        .iter()
-        .any(|allow_rule| domain_rule_matches_host(allow_rule, block_rule))
-}
-
-fn all_blocklist_rules_for_profile(profile: &BlocklistProfileConfig) -> Vec<String> {
-    if profile.categories.is_empty() {
-        return dedup_case_insensitive(profile.sites.iter().cloned());
-    }
-    dedup_case_insensitive(
-        profile
-            .categories
-            .iter()
-            .flat_map(|category| category.sites.iter().cloned()),
-    )
-}
-
-fn all_allowlist_rules_for_profile(profile: &BlocklistProfileConfig) -> Vec<String> {
-    if profile.categories.is_empty() {
-        return dedup_case_insensitive(profile.allowlist_sites.iter().cloned());
-    }
-    dedup_case_insensitive(
-        profile
-            .categories
-            .iter()
-            .flat_map(|category| category.allowlist_sites.iter().cloned()),
-    )
-}
-
-fn dedup_case_insensitive<I>(values: I) -> Vec<String>
-where
-    I: IntoIterator<Item = String>,
-{
-    let mut seen = HashSet::new();
-    values
-        .into_iter()
-        .filter(|value| seen.insert(value.to_ascii_lowercase()))
-        .collect()
 }
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
@@ -1550,118 +1109,6 @@ impl Default for HistoryDashboardConfig {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct WakatimeTaskMappingConfig {
-    #[serde(default)]
-    pub task_label: String,
-    #[serde(default)]
-    pub project: Option<String>,
-    #[serde(default)]
-    pub language: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct WakatimeRuntimeConfig {
-    #[serde(default = "default_wakatime_retry_backoff_secs")]
-    pub retry_backoff_secs: Vec<u64>,
-    #[serde(default = "default_wakatime_queue_capacity")]
-    pub queue_capacity: usize,
-    #[serde(default = "default_wakatime_queue_retry_delay_secs")]
-    pub queue_retry_delay_secs: u64,
-}
-
-impl WakatimeRuntimeConfig {
-    pub fn normalized(&self) -> Self {
-        Self {
-            retry_backoff_secs: normalize_wakatime_retry_backoff_secs(&self.retry_backoff_secs),
-            queue_capacity: self
-                .queue_capacity
-                .clamp(WAKATIME_QUEUE_CAPACITY_MIN, WAKATIME_QUEUE_CAPACITY_MAX),
-            queue_retry_delay_secs: self
-                .queue_retry_delay_secs
-                .clamp(WAKATIME_RETRY_DELAY_MIN_SECS, WAKATIME_RETRY_DELAY_MAX_SECS),
-        }
-    }
-}
-
-impl Default for WakatimeRuntimeConfig {
-    fn default() -> Self {
-        Self {
-            retry_backoff_secs: default_wakatime_retry_backoff_secs(),
-            queue_capacity: default_wakatime_queue_capacity(),
-            queue_retry_delay_secs: default_wakatime_queue_retry_delay_secs(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct WakatimeMetadataConfig {
-    #[serde(default = "default_wakatime_project")]
-    pub project: String,
-    #[serde(default = "default_wakatime_language")]
-    pub language: String,
-    #[serde(default)]
-    pub task_mappings: Vec<WakatimeTaskMappingConfig>,
-}
-
-impl WakatimeMetadataConfig {
-    pub fn normalized(&self) -> Self {
-        Self {
-            project: normalize_nonempty_or_default_string(
-                &self.project,
-                &default_wakatime_project(),
-            ),
-            language: normalize_nonempty_or_default_string(
-                &self.language,
-                &default_wakatime_language(),
-            ),
-            task_mappings: normalize_wakatime_task_mappings(&self.task_mappings),
-        }
-    }
-
-    pub fn task_mapping_for_label(&self, task_label: &str) -> Option<&WakatimeTaskMappingConfig> {
-        let task_label = task_label.trim();
-        if task_label.is_empty() {
-            return None;
-        }
-        self.task_mappings
-            .iter()
-            .find(|mapping| mapping.task_label.eq_ignore_ascii_case(task_label))
-    }
-
-    pub fn resolved_project_language_for_task_label(
-        &self,
-        task_label: Option<&str>,
-    ) -> (String, String) {
-        let Some(task_label) = task_label.map(str::trim).filter(|label| !label.is_empty()) else {
-            return (self.project.clone(), self.language.clone());
-        };
-        let Some(mapping) = self.task_mapping_for_label(task_label) else {
-            return (self.project.clone(), self.language.clone());
-        };
-        (
-            mapping
-                .project
-                .clone()
-                .unwrap_or_else(|| self.project.clone()),
-            mapping
-                .language
-                .clone()
-                .unwrap_or_else(|| self.language.clone()),
-        )
-    }
-}
-
-impl Default for WakatimeMetadataConfig {
-    fn default() -> Self {
-        Self {
-            project: default_wakatime_project(),
-            language: default_wakatime_language(),
-            task_mappings: Vec::new(),
-        }
-    }
-}
-
 impl Default for NotificationConfig {
     fn default() -> Self {
         Self {
@@ -1673,26 +1120,6 @@ impl Default for NotificationConfig {
 
 fn default_notification_enabled() -> bool {
     true
-}
-
-fn default_wakatime_project() -> String {
-    "focustime".to_string()
-}
-
-fn default_wakatime_language() -> String {
-    "Pomodoro".to_string()
-}
-
-fn default_wakatime_retry_backoff_secs() -> Vec<u64> {
-    vec![2, 5, 10]
-}
-
-fn default_wakatime_queue_capacity() -> usize {
-    512
-}
-
-fn default_wakatime_queue_retry_delay_secs() -> u64 {
-    30
 }
 
 fn default_schedule_time_step_minutes() -> u16 {
@@ -1741,212 +1168,12 @@ fn default_one_time_schedule_date() -> String {
     Local::now().date_naive().format("%Y-%m-%d").to_string()
 }
 
-fn default_blocklist_profile_name() -> String {
-    "Default".to_string()
-}
-
-fn default_blocklist_category_name() -> String {
-    "General".to_string()
-}
-
 fn default_break_glass_duration_secs() -> u64 {
     5 * 60
 }
 
 fn default_session_template_name() -> String {
     "Template".to_string()
-}
-
-fn default_shortcut_quit() -> String {
-    "q".to_string()
-}
-
-fn default_shortcut_timer_toggle_pause() -> String {
-    "space".to_string()
-}
-
-fn default_shortcut_timer_stop_reset() -> String {
-    "s".to_string()
-}
-
-fn default_shortcut_timer_next_phase() -> String {
-    "n".to_string()
-}
-
-fn default_shortcut_open_site_manager() -> String {
-    "b".to_string()
-}
-
-fn default_shortcut_open_profile_manager() -> String {
-    "p".to_string()
-}
-
-fn default_shortcut_open_session_planner() -> String {
-    "t".to_string()
-}
-
-fn default_shortcut_open_stats_history() -> String {
-    "h".to_string()
-}
-
-fn default_shortcut_open_setup_diagnostics() -> String {
-    "d".to_string()
-}
-
-fn default_shortcut_timer_edit_note() -> String {
-    "m".to_string()
-}
-
-fn default_shortcut_break_glass_override() -> String {
-    "u".to_string()
-}
-
-fn default_shortcut_delay_schedule_start() -> String {
-    "z".to_string()
-}
-
-fn default_shortcut_back_site_manager() -> String {
-    "b".to_string()
-}
-
-fn default_shortcut_toggle_site_list_mode() -> String {
-    "m".to_string()
-}
-
-fn default_shortcut_site_add() -> String {
-    "a".to_string()
-}
-
-fn default_shortcut_site_edit() -> String {
-    "e".to_string()
-}
-
-fn default_shortcut_site_delete() -> String {
-    "d".to_string()
-}
-
-fn default_shortcut_select_previous_blocklist_profile() -> String {
-    "[".to_string()
-}
-
-fn default_shortcut_select_next_blocklist_profile() -> String {
-    "]".to_string()
-}
-
-fn default_shortcut_create_blocklist_profile() -> String {
-    "n".to_string()
-}
-
-fn default_shortcut_rename_blocklist_profile() -> String {
-    "r".to_string()
-}
-
-fn default_shortcut_delete_blocklist_profile() -> String {
-    "x".to_string()
-}
-
-fn default_shortcut_back_session_planner() -> String {
-    "t".to_string()
-}
-
-fn default_shortcut_planner_add() -> String {
-    "a".to_string()
-}
-
-fn default_shortcut_planner_rename() -> String {
-    "e".to_string()
-}
-
-fn default_shortcut_planner_favorite() -> String {
-    "f".to_string()
-}
-
-fn default_shortcut_planner_archive() -> String {
-    "x".to_string()
-}
-
-fn default_shortcut_planner_delete() -> String {
-    "d".to_string()
-}
-
-fn default_shortcut_planner_select_recent() -> String {
-    "r".to_string()
-}
-
-fn default_shortcut_back_profile_manager() -> String {
-    "p".to_string()
-}
-
-fn default_shortcut_profile_edit() -> String {
-    "e".to_string()
-}
-
-fn default_shortcut_back_stats_history() -> String {
-    "h".to_string()
-}
-
-fn default_shortcut_export_stats_history() -> String {
-    "e".to_string()
-}
-
-fn default_shortcut_history_dashboard_select_previous() -> String {
-    "k".to_string()
-}
-
-fn default_shortcut_history_dashboard_select_next() -> String {
-    "j".to_string()
-}
-
-fn default_shortcut_history_dashboard_toggle_pin() -> String {
-    "p".to_string()
-}
-
-fn default_shortcut_history_dashboard_move_left() -> String {
-    "<".to_string()
-}
-
-fn default_shortcut_history_dashboard_move_right() -> String {
-    ">".to_string()
-}
-
-fn default_shortcut_back_setup_diagnostics() -> String {
-    "d".to_string()
-}
-
-fn default_shortcut_refresh_setup_diagnostics() -> String {
-    "r".to_string()
-}
-
-fn default_shortcut_navigate_up() -> String {
-    "up".to_string()
-}
-
-fn default_shortcut_navigate_down() -> String {
-    "down".to_string()
-}
-
-fn default_shortcut_navigate_left() -> String {
-    "left".to_string()
-}
-
-fn default_shortcut_navigate_right() -> String {
-    "right".to_string()
-}
-
-fn default_shortcut_confirm() -> String {
-    "enter".to_string()
-}
-
-fn default_shortcut_cancel() -> String {
-    "esc".to_string()
-}
-
-fn default_shortcut_delete() -> String {
-    "delete".to_string()
-}
-
-fn default_shortcut_backspace() -> String {
-    "backspace".to_string()
 }
 
 fn default_history_dashboard_pinned_cards() -> Vec<HistoryKpiCardId> {
@@ -2424,877 +1651,6 @@ impl AppConfig {
     }
 }
 
-fn migrate_config_toml_to_current(config_toml: toml::Value) -> Option<toml::Value> {
-    migrate_config_toml_to_current_detailed(config_toml)
-        .ok()
-        .map(|(migrated, _, _)| migrated)
-}
-
-fn migrate_config_toml_to_current_detailed(
-    mut config_toml: toml::Value,
-) -> Result<(toml::Value, u32, Vec<ConfigMigrationStepReport>), String> {
-    let schema_version = detect_config_schema_version(&config_toml)
-        .ok_or_else(|| "Missing or invalid `schema_version` in config.toml.".to_string())?;
-    let mut steps = Vec::new();
-    if schema_version > CURRENT_CONFIG_SCHEMA_VERSION {
-        // Forward-compatibility mode: keep unknown/newer schema content untouched.
-        return Ok((config_toml, schema_version, steps));
-    }
-
-    let mut from_schema_version = schema_version;
-    while from_schema_version < CURRENT_CONFIG_SCHEMA_VERSION {
-        let to_schema_version = from_schema_version + 1;
-        config_toml = migrate_config_toml_step(config_toml, from_schema_version).ok_or_else(|| {
-            format!(
-                "Unsupported migration step from schema v{from_schema_version} to v{to_schema_version}."
-            )
-        })?;
-        steps.push(ConfigMigrationStepReport {
-            from_schema_version,
-            to_schema_version,
-            summary: migration_step_summary(from_schema_version, to_schema_version),
-        });
-        from_schema_version = to_schema_version;
-    }
-    let canonicalization_input = config_toml.clone();
-    canonicalize_legacy_profile_aliases(&mut config_toml);
-    if canonicalization_input != config_toml {
-        steps.push(ConfigMigrationStepReport {
-            from_schema_version,
-            to_schema_version: from_schema_version,
-            summary: "Canonicalize legacy profile aliases in config values.".to_string(),
-        });
-    }
-    Ok((config_toml, schema_version, steps))
-}
-
-fn migration_step_summary(from_schema_version: u32, to_schema_version: u32) -> String {
-    match (from_schema_version, to_schema_version) {
-        (0, 1) => "Add explicit config schema version marker.".to_string(),
-        (1, 2) => {
-            "Rename legacy profile tokens and profile_automation preset keys to canonical values."
-                .to_string()
-        }
-        _ => format!("Migrate config schema from v{from_schema_version} to v{to_schema_version}."),
-    }
-}
-
-fn canonicalize_legacy_profile_aliases(config_toml: &mut toml::Value) {
-    let Some(table) = config_toml.as_table_mut() else {
-        return;
-    };
-    migrate_profile_value_in_table(table, "selected_profile");
-    migrate_profile_automation_preset_keys(table);
-    migrate_profile_value_in_array_table(table, "session_templates", "profile");
-    migrate_profile_value_in_array_table(table, "weekday_profile_rules", "profile");
-    migrate_automation_trigger_action_profiles(table);
-}
-
-fn config_health_warning(
-    code: impl Into<String>,
-    message: impl Into<String>,
-    remediation: impl Into<String>,
-) -> ConfigHealthFinding {
-    ConfigHealthFinding {
-        code: code.into(),
-        severity: ConfigHealthSeverity::Warning,
-        message: message.into(),
-        remediation: remediation.into(),
-    }
-}
-
-fn config_health_error(
-    code: impl Into<String>,
-    message: impl Into<String>,
-    remediation: impl Into<String>,
-) -> ConfigHealthFinding {
-    ConfigHealthFinding {
-        code: code.into(),
-        severity: ConfigHealthSeverity::Error,
-        message: message.into(),
-        remediation: remediation.into(),
-    }
-}
-
-fn summarize_config_health(findings: &[ConfigHealthFinding]) -> ConfigHealthStatus {
-    if findings
-        .iter()
-        .any(|finding| finding.severity == ConfigHealthSeverity::Error)
-    {
-        return ConfigHealthStatus::Error;
-    }
-    if findings.is_empty() {
-        ConfigHealthStatus::Ok
-    } else {
-        ConfigHealthStatus::Warning
-    }
-}
-
-fn sort_config_health_findings(findings: &mut [ConfigHealthFinding]) {
-    findings.sort_by(|left, right| {
-        left.code
-            .cmp(&right.code)
-            .then_with(|| left.message.cmp(&right.message))
-    });
-}
-
-fn legacy_profile_token_migration_target(value: &str) -> Option<&'static str> {
-    match value.trim().to_ascii_lowercase().as_str() {
-        "classic" => Some("basic"),
-        "deep-work" | "deep_work" | "deepwork" => Some("standard"),
-        "custom" => Some("advanced"),
-        _ => None,
-    }
-}
-
-fn collect_legacy_profile_rename_advice(config_toml: &toml::Value) -> Vec<String> {
-    let Some(table) = config_toml.as_table() else {
-        return Vec::new();
-    };
-
-    let mut advice = Vec::new();
-    if let Some(value) = table.get("selected_profile").and_then(toml::Value::as_str) {
-        push_legacy_profile_value_advice(&mut advice, "selected_profile", value);
-    }
-
-    push_legacy_profile_value_array_advice(
-        &mut advice,
-        table,
-        "session_templates",
-        "profile",
-        "session_templates",
-    );
-    push_legacy_profile_value_array_advice(
-        &mut advice,
-        table,
-        "weekday_profile_rules",
-        "profile",
-        "weekday_profile_rules",
-    );
-    push_legacy_automation_trigger_profile_advice(&mut advice, table);
-    push_legacy_profile_automation_key_advice(&mut advice, table);
-
-    advice.sort_unstable();
-    advice.dedup();
-    advice
-}
-
-fn push_legacy_profile_value_array_advice(
-    advice: &mut Vec<String>,
-    table: &toml::map::Map<String, toml::Value>,
-    array_key: &str,
-    field_key: &str,
-    location_prefix: &str,
-) {
-    let Some(array) = table.get(array_key).and_then(toml::Value::as_array) else {
-        return;
-    };
-    for (index, entry) in array.iter().enumerate() {
-        let Some(entry_table) = entry.as_table() else {
-            continue;
-        };
-        let Some(value) = entry_table.get(field_key).and_then(toml::Value::as_str) else {
-            continue;
-        };
-        push_legacy_profile_value_advice(
-            advice,
-            &format!("{location_prefix}[{index}].{field_key}"),
-            value,
-        );
-    }
-}
-
-fn push_legacy_automation_trigger_profile_advice(
-    advice: &mut Vec<String>,
-    table: &toml::map::Map<String, toml::Value>,
-) {
-    let Some(array) = table
-        .get("automation_triggers")
-        .and_then(toml::Value::as_array)
-    else {
-        return;
-    };
-    for (index, entry) in array.iter().enumerate() {
-        let Some(entry_table) = entry.as_table() else {
-            continue;
-        };
-        let Some(action_table) = entry_table.get("action").and_then(toml::Value::as_table) else {
-            continue;
-        };
-        let Some(value) = action_table.get("profile").and_then(toml::Value::as_str) else {
-            continue;
-        };
-        push_legacy_profile_value_advice(
-            advice,
-            &format!("automation_triggers[{index}].action.profile"),
-            value,
-        );
-    }
-}
-
-fn push_legacy_profile_automation_key_advice(
-    advice: &mut Vec<String>,
-    table: &toml::map::Map<String, toml::Value>,
-) {
-    let Some(profile_automation) = table
-        .get("profile_automation")
-        .and_then(toml::Value::as_table)
-    else {
-        return;
-    };
-
-    for (legacy_key, canonical_key) in [
-        ("classic", "basic"),
-        ("deep_work", "standard"),
-        ("deep-work", "standard"),
-        ("custom", "advanced"),
-    ] {
-        if profile_automation.contains_key(legacy_key) {
-            advice.push(format!(
-                "[profile_automation.{legacy_key}] should be renamed to [profile_automation.{canonical_key}]."
-            ));
-        }
-    }
-}
-
-fn push_legacy_profile_value_advice(advice: &mut Vec<String>, location: &str, value: &str) {
-    let Some(target) = legacy_profile_token_migration_target(value) else {
-        return;
-    };
-    advice.push(format!(
-        "{location} uses legacy value \"{value}\"; migrate it to \"{target}\"."
-    ));
-}
-
-fn next_config_backup_path(path: &Path) -> PathBuf {
-    let mut index: usize = 0;
-    loop {
-        let candidate = if index == 0 {
-            path.with_extension("toml.bak")
-        } else {
-            path.with_extension(format!("toml.bak.{index}"))
-        };
-        if !candidate.exists() {
-            return candidate;
-        }
-        index += 1;
-    }
-}
-
-fn write_atomic_text_file(path: &Path, content: &str) -> io::Result<()> {
-    if let Some(parent) = path.parent() {
-        fs::create_dir_all(parent)?;
-    }
-    let tmp = path.with_extension("toml.tmp");
-    fs::write(&tmp, content)?;
-    #[cfg(target_os = "windows")]
-    {
-        match fs::rename(&tmp, path) {
-            Ok(()) => Ok(()),
-            Err(error) if error.kind() == io::ErrorKind::AlreadyExists => {
-                fs::remove_file(path)?;
-                fs::rename(&tmp, path)
-            }
-            Err(error) => Err(error),
-        }
-    }
-    #[cfg(not(target_os = "windows"))]
-    {
-        fs::rename(&tmp, path)
-    }
-}
-
-fn write_migrated_config_with_backup(
-    path: &Path,
-    migrated_toml: &toml::Value,
-) -> Result<PathBuf, String> {
-    let backup_path = next_config_backup_path(path);
-    fs::copy(path, &backup_path).map_err(|error| {
-        format!(
-            "Failed to create config backup at `{}`: {error}",
-            backup_path.display()
-        )
-    })?;
-    let content = toml::to_string_pretty(migrated_toml)
-        .map_err(|error| format!("Failed to serialize migrated config TOML: {error}"))?;
-    write_atomic_text_file(path, &content).map_err(|error| {
-        format!(
-            "Failed to write migrated config to `{}`: {error}",
-            path.display()
-        )
-    })?;
-    Ok(backup_path)
-}
-
-pub fn run_config_doctor() -> ConfigDoctorReport {
-    run_config_doctor_with_path(AppConfig::config_path())
-}
-
-fn run_config_doctor_with_path(config_path: Option<PathBuf>) -> ConfigDoctorReport {
-    let action = "config-doctor";
-    let mut detected_schema_version = None;
-    let mut migration_steps = Vec::new();
-    let mut findings = Vec::new();
-
-    let Some(path) = config_path.clone() else {
-        findings.push(config_health_error(
-            "config.path_unavailable",
-            "Could not determine the config.toml path from the current environment.",
-            "Set the platform config directory environment variables (for example APPDATA on Windows, XDG_CONFIG_HOME/HOME on Unix), then retry.",
-        ));
-        return ConfigDoctorReport {
-            action,
-            config_path,
-            detected_schema_version,
-            current_schema_version: CURRENT_CONFIG_SCHEMA_VERSION,
-            status: summarize_config_health(&findings),
-            migration_steps,
-            findings,
-        };
-    };
-
-    if !path.exists() {
-        findings.push(config_health_warning(
-            "config.file_missing",
-            format!("Config file `{}` does not exist.", path.display()),
-            "Run focustime once (or create config.toml), then rerun the doctor.",
-        ));
-        return ConfigDoctorReport {
-            action,
-            config_path: Some(path),
-            detected_schema_version,
-            current_schema_version: CURRENT_CONFIG_SCHEMA_VERSION,
-            status: summarize_config_health(&findings),
-            migration_steps,
-            findings,
-        };
-    }
-
-    let content = match fs::read_to_string(&path) {
-        Ok(content) => content,
-        Err(error) => {
-            findings.push(config_health_error(
-                "config.read_failed",
-                format!("Failed to read `{}`: {error}", path.display()),
-                "Check filesystem permissions and path accessibility, then rerun the doctor.",
-            ));
-            return ConfigDoctorReport {
-                action,
-                config_path: Some(path),
-                detected_schema_version,
-                current_schema_version: CURRENT_CONFIG_SCHEMA_VERSION,
-                status: summarize_config_health(&findings),
-                migration_steps,
-                findings,
-            };
-        }
-    };
-
-    let original_toml: toml::Value = match toml::from_str(&content) {
-        Ok(value) => value,
-        Err(error) => {
-            findings.push(config_health_error(
-                "config.toml_parse_error",
-                format!("config.toml is not valid TOML: {error}"),
-                "Fix the TOML syntax error and rerun the doctor.",
-            ));
-            return ConfigDoctorReport {
-                action,
-                config_path: Some(path),
-                detected_schema_version,
-                current_schema_version: CURRENT_CONFIG_SCHEMA_VERSION,
-                status: summarize_config_health(&findings),
-                migration_steps,
-                findings,
-            };
-        }
-    };
-
-    let rename_advice = collect_legacy_profile_rename_advice(&original_toml);
-    let (migrated_toml, schema_version, steps) =
-        match migrate_config_toml_to_current_detailed(original_toml.clone()) {
-            Ok(result) => result,
-            Err(error) => {
-                findings.push(config_health_error(
-                    "config.migration_failed",
-                    format!("Config migration analysis failed: {error}"),
-                    "Ensure schema_version is valid and the file is structurally correct TOML.",
-                ));
-                return ConfigDoctorReport {
-                    action,
-                    config_path: Some(path),
-                    detected_schema_version,
-                    current_schema_version: CURRENT_CONFIG_SCHEMA_VERSION,
-                    status: summarize_config_health(&findings),
-                    migration_steps,
-                    findings,
-                };
-            }
-        };
-    detected_schema_version = Some(schema_version);
-    migration_steps = steps;
-
-    if schema_version > CURRENT_CONFIG_SCHEMA_VERSION {
-        findings.push(config_health_warning(
-            "config.schema_newer_than_supported",
-            format!(
-                "Config schema version {schema_version} is newer than this build's supported schema version {}.",
-                CURRENT_CONFIG_SCHEMA_VERSION
-            ),
-            "Use a focustime build that supports this schema before writing config changes.",
-        ));
-        sort_config_health_findings(&mut findings);
-        return ConfigDoctorReport {
-            action,
-            config_path: Some(path),
-            detected_schema_version,
-            current_schema_version: CURRENT_CONFIG_SCHEMA_VERSION,
-            status: summarize_config_health(&findings),
-            migration_steps,
-            findings,
-        };
-    } else if schema_version < CURRENT_CONFIG_SCHEMA_VERSION {
-        findings.push(config_health_warning(
-            "config.schema_outdated",
-            format!(
-                "Config schema version {schema_version} is older than current schema version {}.",
-                CURRENT_CONFIG_SCHEMA_VERSION
-            ),
-            "Run `focustime --config-migrate` to preview migration and `focustime --config-migrate-apply` to apply it.",
-        ));
-    }
-
-    for advice in rename_advice {
-        findings.push(config_health_warning(
-            "config.legacy_profile_token",
-            advice,
-            "Run `focustime --config-migrate` to preview canonical profile-key migrations.",
-        ));
-    }
-
-    let disk: AppConfigDisk = match migrated_toml.try_into() {
-        Ok(disk) => disk,
-        Err(error) => {
-            findings.push(config_health_error(
-                "config.deserialize_failed",
-                format!("Config could not be deserialized after migration steps: {error}"),
-                "Fix incompatible field types/structures in config.toml, then rerun the doctor.",
-            ));
-            return ConfigDoctorReport {
-                action,
-                config_path: Some(path),
-                detected_schema_version,
-                current_schema_version: CURRENT_CONFIG_SCHEMA_VERSION,
-                status: summarize_config_health(&findings),
-                migration_steps,
-                findings,
-            };
-        }
-    };
-
-    let config = disk.config;
-    for warning in detect_legacy_config_deprecation_warnings(&config) {
-        findings.push(config_health_warning(
-            "config.deprecated_field_in_use",
-            warning,
-            "Update config.toml to canonical fields and rerun the doctor.",
-        ));
-    }
-    let normalized = config.normalize();
-    if let Err(error) = validate_automation_trigger_rules(
-        &normalized.automation_triggers,
-        &normalized.blocklist_profiles,
-        &normalized.session_templates,
-    ) {
-        findings.push(config_health_error(
-            "config.automation_trigger_conflict",
-            format!("Automation trigger configuration conflict: {error}"),
-            "Fix the conflicting trigger rules and rerun the doctor.",
-        ));
-    }
-
-    sort_config_health_findings(&mut findings);
-    ConfigDoctorReport {
-        action,
-        config_path: Some(path),
-        detected_schema_version,
-        current_schema_version: CURRENT_CONFIG_SCHEMA_VERSION,
-        status: summarize_config_health(&findings),
-        migration_steps,
-        findings,
-    }
-}
-
-pub fn run_config_migration_assistant(apply: bool) -> ConfigMigrationReport {
-    run_config_migration_assistant_with_path(apply, AppConfig::config_path())
-}
-
-fn run_config_migration_assistant_with_path(
-    apply: bool,
-    config_path: Option<PathBuf>,
-) -> ConfigMigrationReport {
-    let action = if apply {
-        "config-migrate-apply"
-    } else {
-        "config-migrate"
-    };
-    let mut detected_schema_version = None;
-    let mut backup_path = None;
-    let mut steps = Vec::new();
-    let mut changed = false;
-    let mut applied = false;
-    let mut findings = Vec::new();
-
-    let Some(path) = config_path.clone() else {
-        findings.push(config_health_error(
-            "config.path_unavailable",
-            "Could not determine the config.toml path from the current environment.",
-            "Set the platform config directory environment variables (for example APPDATA on Windows, XDG_CONFIG_HOME/HOME on Unix), then retry.",
-        ));
-        return ConfigMigrationReport {
-            action,
-            applied,
-            config_path,
-            backup_path,
-            detected_schema_version,
-            target_schema_version: CURRENT_CONFIG_SCHEMA_VERSION,
-            changed,
-            status: summarize_config_health(&findings),
-            steps,
-            findings,
-        };
-    };
-
-    if !path.exists() {
-        findings.push(config_health_warning(
-            "config.file_missing",
-            format!("Config file `{}` does not exist.", path.display()),
-            "Run focustime once (or create config.toml), then rerun config migration.",
-        ));
-        return ConfigMigrationReport {
-            action,
-            applied,
-            config_path: Some(path),
-            backup_path,
-            detected_schema_version,
-            target_schema_version: CURRENT_CONFIG_SCHEMA_VERSION,
-            changed,
-            status: summarize_config_health(&findings),
-            steps,
-            findings,
-        };
-    }
-
-    let content = match fs::read_to_string(&path) {
-        Ok(content) => content,
-        Err(error) => {
-            findings.push(config_health_error(
-                "config.read_failed",
-                format!("Failed to read `{}`: {error}", path.display()),
-                "Check filesystem permissions and path accessibility, then rerun migration.",
-            ));
-            return ConfigMigrationReport {
-                action,
-                applied,
-                config_path: Some(path),
-                backup_path,
-                detected_schema_version,
-                target_schema_version: CURRENT_CONFIG_SCHEMA_VERSION,
-                changed,
-                status: summarize_config_health(&findings),
-                steps,
-                findings,
-            };
-        }
-    };
-
-    let original_toml: toml::Value = match toml::from_str(&content) {
-        Ok(value) => value,
-        Err(error) => {
-            findings.push(config_health_error(
-                "config.toml_parse_error",
-                format!("config.toml is not valid TOML: {error}"),
-                "Fix the TOML syntax error and rerun config migration.",
-            ));
-            return ConfigMigrationReport {
-                action,
-                applied,
-                config_path: Some(path),
-                backup_path,
-                detected_schema_version,
-                target_schema_version: CURRENT_CONFIG_SCHEMA_VERSION,
-                changed,
-                status: summarize_config_health(&findings),
-                steps,
-                findings,
-            };
-        }
-    };
-
-    for advice in collect_legacy_profile_rename_advice(&original_toml) {
-        findings.push(config_health_warning(
-            "config.legacy_profile_token",
-            advice,
-            "Use canonical profile tokens (`basic`, `standard`, `advanced`).",
-        ));
-    }
-
-    let (migrated_toml, schema_version, migration_steps) =
-        match migrate_config_toml_to_current_detailed(original_toml.clone()) {
-            Ok(result) => result,
-            Err(error) => {
-                findings.push(config_health_error(
-                    "config.migration_failed",
-                    format!("Migration analysis failed: {error}"),
-                    "Ensure schema_version is valid and the file is structurally correct TOML.",
-                ));
-                return ConfigMigrationReport {
-                    action,
-                    applied,
-                    config_path: Some(path),
-                    backup_path,
-                    detected_schema_version,
-                    target_schema_version: CURRENT_CONFIG_SCHEMA_VERSION,
-                    changed,
-                    status: summarize_config_health(&findings),
-                    steps,
-                    findings,
-                };
-            }
-        };
-    detected_schema_version = Some(schema_version);
-    steps = migration_steps;
-    changed = original_toml != migrated_toml;
-
-    if schema_version > CURRENT_CONFIG_SCHEMA_VERSION {
-        findings.push(config_health_warning(
-            "config.schema_newer_than_supported",
-            format!(
-                "Config schema version {schema_version} is newer than this build's supported schema version {}.",
-                CURRENT_CONFIG_SCHEMA_VERSION
-            ),
-            "Use a focustime build that supports this schema before applying migration.",
-        ));
-    } else if !changed {
-        findings.push(config_health_warning(
-            "config.already_current",
-            "Config already matches the current migration schema and canonical key mapping."
-                .to_string(),
-            "No migration apply step is required.",
-        ));
-    }
-
-    if apply && changed && schema_version <= CURRENT_CONFIG_SCHEMA_VERSION {
-        match write_migrated_config_with_backup(&path, &migrated_toml) {
-            Ok(created_backup) => {
-                backup_path = Some(created_backup);
-                applied = true;
-            }
-            Err(error) => {
-                findings.push(config_health_error(
-                    "config.apply_failed",
-                    error,
-                    "Fix the file system error, restore from backup if needed, and retry migration apply.",
-                ));
-            }
-        }
-    }
-
-    sort_config_health_findings(&mut findings);
-    ConfigMigrationReport {
-        action,
-        applied,
-        config_path: Some(path),
-        backup_path,
-        detected_schema_version,
-        target_schema_version: CURRENT_CONFIG_SCHEMA_VERSION,
-        changed,
-        status: summarize_config_health(&findings),
-        steps,
-        findings,
-    }
-}
-
-fn detect_legacy_config_deprecation_warnings(config: &AppConfig) -> Vec<String> {
-    let mut warnings = Vec::new();
-    let duration_override_without_custom_profile = config.custom_profile.is_none()
-        && (config.focus_secs != default_focus_secs()
-            || config.short_break_secs != default_short_break_secs()
-            || config.long_break_secs != default_long_break_secs()
-            || config.long_break_interval != default_long_break_interval());
-    if duration_override_without_custom_profile {
-        warnings.push(
-            "Deprecated top-level timer fields (`focus_secs`, `short_break_secs`, `long_break_secs`, `long_break_interval`) are in use. Move these values into `[custom_profile]`.".to_string(),
-        );
-    }
-
-    let legacy_automation_values_configured = config.notifications != NotificationConfig::default()
-        || config.auto_start != AutoStartConfig::default()
-        || config.strict_mode
-        || config.recurring_schedule != RecurringScheduleConfig::default();
-    let profile_automation_incomplete = config
-        .profile_automation
-        .as_ref()
-        .map(|settings| {
-            settings.basic.is_none() || settings.standard.is_none() || settings.advanced.is_none()
-        })
-        .unwrap_or(true);
-    let legacy_automation_in_use =
-        legacy_automation_values_configured && profile_automation_incomplete;
-    if legacy_automation_in_use {
-        warnings.push(
-            "Deprecated top-level automation fields (`notifications`, `auto_start`, `strict_mode`, `recurring_schedule`) are in use. Move them under `[profile_automation.<preset>]`.".to_string(),
-        );
-    }
-
-    if config.blocklist_profiles.is_empty() && !config.blocked_sites.is_empty() {
-        warnings.push(
-            "Deprecated `blocked_sites` is in use without `[[blocklist_profiles]]`. Move entries into a blocklist profile (for example `Default`).".to_string(),
-        );
-    }
-
-    warnings
-}
-
-fn detect_config_schema_version(config_toml: &toml::Value) -> Option<u32> {
-    let table = config_toml.as_table()?;
-    table
-        .get("schema_version")
-        .map(|value| value.as_integer().and_then(|raw| u32::try_from(raw).ok()))
-        .unwrap_or(Some(LEGACY_CONFIG_SCHEMA_VERSION))
-}
-
-fn migrate_config_toml_step(
-    config_toml: toml::Value,
-    from_schema_version: u32,
-) -> Option<toml::Value> {
-    match from_schema_version {
-        LEGACY_CONFIG_SCHEMA_VERSION => migrate_config_toml_legacy_to_v1(config_toml),
-        1 => migrate_config_toml_v1_to_v2(config_toml),
-        _ => None,
-    }
-}
-
-fn migrate_config_toml_legacy_to_v1(mut config_toml: toml::Value) -> Option<toml::Value> {
-    let table = config_toml.as_table_mut()?;
-    table.insert("schema_version".to_string(), toml::Value::Integer(1));
-    Some(config_toml)
-}
-
-fn migrate_config_toml_v1_to_v2(mut config_toml: toml::Value) -> Option<toml::Value> {
-    let table = config_toml.as_table_mut()?;
-    migrate_profile_value_in_table(table, "selected_profile");
-    migrate_profile_automation_preset_keys(table);
-    migrate_profile_value_in_array_table(table, "session_templates", "profile");
-    migrate_profile_value_in_array_table(table, "weekday_profile_rules", "profile");
-    migrate_automation_trigger_action_profiles(table);
-    table.insert(
-        "schema_version".to_string(),
-        toml::Value::Integer(i64::from(CURRENT_CONFIG_SCHEMA_VERSION)),
-    );
-    Some(config_toml)
-}
-
-fn migrate_profile_value_in_array_table(
-    table: &mut toml::map::Map<String, toml::Value>,
-    array_key: &str,
-    field_key: &str,
-) {
-    let Some(array) = table
-        .get_mut(array_key)
-        .and_then(|value| value.as_array_mut())
-    else {
-        return;
-    };
-    for entry in array {
-        let Some(entry_table) = entry.as_table_mut() else {
-            continue;
-        };
-        migrate_profile_value_in_table(entry_table, field_key);
-    }
-}
-
-fn migrate_profile_value_in_table(
-    table: &mut toml::map::Map<String, toml::Value>,
-    field_key: &str,
-) {
-    let Some(value) = table.get_mut(field_key) else {
-        return;
-    };
-    let Some(raw) = value.as_str() else {
-        return;
-    };
-    let Some(mapped) = canonical_profile_id_token(raw) else {
-        return;
-    };
-    *value = toml::Value::String(mapped.to_string());
-}
-
-fn migrate_profile_automation_preset_keys(table: &mut toml::map::Map<String, toml::Value>) {
-    let Some(profile_automation) = table
-        .get_mut("profile_automation")
-        .and_then(|value| value.as_table_mut())
-    else {
-        return;
-    };
-    migrate_table_key(profile_automation, "classic", "basic");
-    migrate_table_key(profile_automation, "deep_work", "standard");
-    migrate_table_key(profile_automation, "deep-work", "standard");
-    migrate_table_key(profile_automation, "custom", "advanced");
-}
-
-fn migrate_table_key(
-    table: &mut toml::map::Map<String, toml::Value>,
-    old_key: &str,
-    new_key: &str,
-) {
-    let Some(value) = table.remove(old_key) else {
-        return;
-    };
-    if let Some(existing) = table.get_mut(new_key) {
-        merge_toml_value_prefer_existing(existing, value);
-        return;
-    }
-    table.insert(new_key.to_string(), value);
-}
-
-fn merge_toml_value_prefer_existing(existing: &mut toml::Value, incoming: toml::Value) {
-    let (toml::Value::Table(existing_table), toml::Value::Table(incoming_table)) =
-        (existing, incoming)
-    else {
-        return;
-    };
-
-    for (key, incoming_value) in incoming_table {
-        if let Some(existing_value) = existing_table.get_mut(&key) {
-            merge_toml_value_prefer_existing(existing_value, incoming_value);
-        } else {
-            existing_table.insert(key, incoming_value);
-        }
-    }
-}
-
-fn migrate_automation_trigger_action_profiles(table: &mut toml::map::Map<String, toml::Value>) {
-    let Some(triggers) = table
-        .get_mut("automation_triggers")
-        .and_then(|value| value.as_array_mut())
-    else {
-        return;
-    };
-    for entry in triggers {
-        let Some(entry_table) = entry.as_table_mut() else {
-            continue;
-        };
-        let Some(action_table) = entry_table
-            .get_mut("action")
-            .and_then(|action| action.as_table_mut())
-        else {
-            continue;
-        };
-        migrate_profile_value_in_table(action_table, "profile");
-    }
-}
-
 #[cfg_attr(test, allow(dead_code))]
 pub(crate) fn app_data_path(file_name: &str) -> Option<PathBuf> {
     let app_dir = app_dir()?;
@@ -3341,54 +1697,6 @@ fn normalize_optional_nonempty_string(value: Option<&str>) -> Option<String> {
     }
 }
 
-fn normalize_wakatime_task_mappings(
-    mappings: &[WakatimeTaskMappingConfig],
-) -> Vec<WakatimeTaskMappingConfig> {
-    let mut normalized = Vec::new();
-    let mut seen_labels = HashSet::new();
-    for mapping in mappings {
-        let Some(task_label) = normalize_optional_nonempty_string(Some(&mapping.task_label)) else {
-            continue;
-        };
-        let project = normalize_optional_nonempty_string(mapping.project.as_deref());
-        let language = normalize_optional_nonempty_string(mapping.language.as_deref());
-        if project.is_none() && language.is_none() {
-            continue;
-        }
-        let key = task_label.to_ascii_lowercase();
-        if seen_labels.insert(key) {
-            normalized.push(WakatimeTaskMappingConfig {
-                task_label,
-                project,
-                language,
-            });
-        }
-    }
-    normalized
-}
-
-fn normalize_wakatime_retry_backoff_secs(backoff_secs: &[u64]) -> Vec<u64> {
-    let normalized = backoff_secs
-        .iter()
-        .filter_map(|secs| {
-            if *secs == 0 {
-                None
-            } else {
-                Some((*secs).clamp(
-                    WAKATIME_RETRY_BACKOFF_MIN_SECS,
-                    WAKATIME_RETRY_BACKOFF_MAX_SECS,
-                ))
-            }
-        })
-        .take(WAKATIME_RETRY_BACKOFF_MAX_ENTRIES)
-        .collect::<Vec<_>>();
-    if normalized.is_empty() {
-        default_wakatime_retry_backoff_secs()
-    } else {
-        normalized
-    }
-}
-
 fn normalize_calendar_sources(sources: &[CalendarSourceConfig]) -> Vec<CalendarSourceConfig> {
     let mut normalized = Vec::new();
     let mut seen_keys = HashSet::new();
@@ -3412,90 +1720,6 @@ fn normalize_calendar_sources(sources: &[CalendarSourceConfig]) -> Vec<CalendarS
         }
     }
     normalized
-}
-
-fn parse_shortcut_char(value: &str) -> Option<char> {
-    let token = value.trim();
-    if token.eq_ignore_ascii_case("space") {
-        return Some(' ');
-    }
-
-    let mut chars = token.chars();
-    let candidate = chars.next()?;
-    if chars.next().is_some() {
-        return None;
-    }
-
-    if candidate.is_ascii_alphabetic() {
-        Some(candidate.to_ascii_lowercase())
-    } else {
-        Some(candidate)
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ShortcutKeyToken {
-    Char(char),
-    Enter,
-    Esc,
-    Up,
-    Down,
-    Left,
-    Right,
-    Backspace,
-    Delete,
-}
-
-fn parse_shortcut_key_token(value: &str) -> Option<ShortcutKeyToken> {
-    let token = value.trim();
-    if token.is_empty() {
-        return None;
-    }
-
-    match token.to_ascii_lowercase().as_str() {
-        "enter" => Some(ShortcutKeyToken::Enter),
-        "esc" | "escape" => Some(ShortcutKeyToken::Esc),
-        "up" => Some(ShortcutKeyToken::Up),
-        "down" => Some(ShortcutKeyToken::Down),
-        "left" => Some(ShortcutKeyToken::Left),
-        "right" => Some(ShortcutKeyToken::Right),
-        "backspace" => Some(ShortcutKeyToken::Backspace),
-        "delete" | "del" => Some(ShortcutKeyToken::Delete),
-        _ => parse_shortcut_char(token).map(ShortcutKeyToken::Char),
-    }
-}
-
-fn normalize_shortcut_token(value: &str, default: &str) -> String {
-    let Some(parsed) = parse_shortcut_char(value) else {
-        return default.to_string();
-    };
-    if parsed == ' ' {
-        "space".to_string()
-    } else if parsed.is_ascii_alphabetic() {
-        parsed.to_ascii_lowercase().to_string()
-    } else {
-        parsed.to_string()
-    }
-}
-
-fn normalize_navigation_shortcut_token(value: &str, default: &str) -> String {
-    let Some(parsed) = parse_shortcut_key_token(value) else {
-        return default.to_string();
-    };
-
-    match parsed {
-        ShortcutKeyToken::Char(' ') => "space".to_string(),
-        ShortcutKeyToken::Char(c) if c.is_ascii_alphabetic() => c.to_ascii_lowercase().to_string(),
-        ShortcutKeyToken::Char(c) => c.to_string(),
-        ShortcutKeyToken::Enter => "enter".to_string(),
-        ShortcutKeyToken::Esc => "esc".to_string(),
-        ShortcutKeyToken::Up => "up".to_string(),
-        ShortcutKeyToken::Down => "down".to_string(),
-        ShortcutKeyToken::Left => "left".to_string(),
-        ShortcutKeyToken::Right => "right".to_string(),
-        ShortcutKeyToken::Backspace => "backspace".to_string(),
-        ShortcutKeyToken::Delete => "delete".to_string(),
-    }
 }
 
 fn normalize_schedule_days(days: &[String]) -> Vec<String> {
@@ -3601,512 +1825,6 @@ fn normalize_optional_selected_session_template(
         .iter()
         .find(|template| template.name.eq_ignore_ascii_case(&normalized))
         .map(|template| template.name.clone())
-}
-
-fn normalize_automation_triggers(
-    rules: &[AutomationTriggerRuleConfig],
-    blocklist_profiles: &[BlocklistProfileConfig],
-    session_templates: &[SessionTemplateConfig],
-) -> Vec<AutomationTriggerRuleConfig> {
-    rules
-        .iter()
-        .filter_map(|rule| rule.normalized_with_context(blocklist_profiles, session_templates))
-        .collect()
-}
-
-pub fn validate_automation_trigger_rules(
-    rules: &[AutomationTriggerRuleConfig],
-    blocklist_profiles: &[BlocklistProfileConfig],
-    session_templates: &[SessionTemplateConfig],
-) -> Result<(), String> {
-    let mut seen_trigger_keys: HashMap<String, AutomationTriggerConflictRule> = HashMap::new();
-    for (index, rule) in rules.iter().enumerate() {
-        validate_automation_trigger_rule(
-            rule,
-            index,
-            blocklist_profiles,
-            session_templates,
-            &mut seen_trigger_keys,
-        )?;
-    }
-    Ok(())
-}
-
-fn validate_automation_trigger_rule(
-    rule: &AutomationTriggerRuleConfig,
-    index: usize,
-    blocklist_profiles: &[BlocklistProfileConfig],
-    session_templates: &[SessionTemplateConfig],
-    seen_trigger_keys: &mut HashMap<String, AutomationTriggerConflictRule>,
-) -> Result<(), String> {
-    validate_automation_trigger_condition(&rule.trigger, index)?;
-    validate_automation_trigger_action(&rule.action, index, blocklist_profiles, session_templates)?;
-    validate_automation_trigger_conflicts(&rule.trigger, &rule.action, index, seen_trigger_keys)?;
-    Ok(())
-}
-
-fn validate_automation_trigger_condition(
-    trigger: &AutomationTriggerConditionConfig,
-    index: usize,
-) -> Result<(), String> {
-    match trigger {
-        AutomationTriggerConditionConfig::ScheduleWindowStart
-        | AutomationTriggerConditionConfig::ScheduleWindowEnd
-        | AutomationTriggerConditionConfig::FocusStarted
-        | AutomationTriggerConditionConfig::FocusCompleted
-        | AutomationTriggerConditionConfig::BreakStarted
-        | AutomationTriggerConditionConfig::BreakCompleted => Ok(()),
-        AutomationTriggerConditionConfig::Time { days, at } => {
-            if days.is_empty() {
-                return Err(format!(
-                    "Invalid automation trigger rule at index {index}: time trigger `days` cannot be empty."
-                ));
-            }
-            for day in days {
-                if weekday_token_to_index(day).is_none() {
-                    return Err(format!(
-                        "Invalid automation trigger rule at index {index}: unknown weekday `{day}` in time trigger."
-                    ));
-                }
-            }
-            if parse_schedule_time_minutes(at).is_none() {
-                return Err(format!(
-                    "Invalid automation trigger rule at index {index}: time trigger `at` must be HH:MM in 24-hour format."
-                ));
-            }
-            Ok(())
-        }
-    }
-}
-
-fn validate_automation_trigger_action(
-    action: &AutomationTriggerActionConfig,
-    index: usize,
-    blocklist_profiles: &[BlocklistProfileConfig],
-    session_templates: &[SessionTemplateConfig],
-) -> Result<(), String> {
-    match action {
-        AutomationTriggerActionConfig::StartFocus => Ok(()),
-        AutomationTriggerActionConfig::DelayScheduleStart { delay_secs } => {
-            if *delay_secs < SCHEDULE_DELAY_MIN_SECS || *delay_secs > SCHEDULE_DELAY_MAX_SECS {
-                return Err(format!(
-                    "Invalid automation trigger rule at index {index}: `delay_secs` must be between {SCHEDULE_DELAY_MIN_SECS} and {SCHEDULE_DELAY_MAX_SECS}."
-                ));
-            }
-            Ok(())
-        }
-        AutomationTriggerActionConfig::ApplyDefaults {
-            blocklist_profile,
-            session_template,
-            ..
-        } => validate_automation_trigger_apply_defaults_action(
-            blocklist_profile,
-            session_template.as_deref(),
-            index,
-            blocklist_profiles,
-            session_templates,
-        ),
-    }
-}
-
-fn validate_automation_trigger_apply_defaults_action(
-    blocklist_profile: &str,
-    session_template: Option<&str>,
-    index: usize,
-    blocklist_profiles: &[BlocklistProfileConfig],
-    session_templates: &[SessionTemplateConfig],
-) -> Result<(), String> {
-    if blocklist_profile.trim().is_empty() {
-        return Err(format!(
-            "Invalid automation trigger rule at index {index}: `blocklist_profile` cannot be empty."
-        ));
-    }
-    if !blocklist_profiles
-        .iter()
-        .any(|profile| profile.name.eq_ignore_ascii_case(blocklist_profile.trim()))
-    {
-        return Err(format!(
-            "Invalid automation trigger rule at index {index}: blocklist profile `{}` does not exist.",
-            blocklist_profile
-        ));
-    }
-    if let Some(template) = session_template {
-        if template.trim().is_empty() {
-            return Err(format!(
-                "Invalid automation trigger rule at index {index}: `session_template` cannot be empty when provided."
-            ));
-        }
-        if !session_templates
-            .iter()
-            .any(|candidate| candidate.name.eq_ignore_ascii_case(template.trim()))
-        {
-            return Err(format!(
-                "Invalid automation trigger rule at index {index}: session template `{template}` does not exist."
-            ));
-        }
-    }
-    Ok(())
-}
-
-fn validate_automation_trigger_conflicts(
-    trigger: &AutomationTriggerConditionConfig,
-    action: &AutomationTriggerActionConfig,
-    index: usize,
-    seen_trigger_keys: &mut HashMap<String, AutomationTriggerConflictRule>,
-) -> Result<(), String> {
-    let conflict_keys = automation_trigger_conflict_keys(trigger);
-    for trigger_key in conflict_keys {
-        if let Some(previous_rule) = seen_trigger_keys.get(&trigger_key)
-            && automation_trigger_actions_conflict(&previous_rule.action, action)
-        {
-            let previous_action = format_automation_trigger_action(&previous_rule.action);
-            let current_action = format_automation_trigger_action(action);
-            return Err(format!(
-                "Conflicting automation trigger rules: rule #{} (`{previous_action}`) conflicts with rule #{} (`{current_action}`) because both target {}. Keep one rule per trigger condition, or change one trigger so they do not overlap.",
-                previous_rule.index + 1,
-                index + 1,
-                format_automation_trigger_conflict_key(&trigger_key),
-            ));
-        }
-        seen_trigger_keys.insert(
-            trigger_key,
-            AutomationTriggerConflictRule {
-                index,
-                action: action.clone(),
-            },
-        );
-    }
-    Ok(())
-}
-
-fn automation_trigger_conflict_keys(trigger: &AutomationTriggerConditionConfig) -> Vec<String> {
-    match trigger {
-        AutomationTriggerConditionConfig::ScheduleWindowStart => {
-            vec!["schedule_window_start".to_string()]
-        }
-        AutomationTriggerConditionConfig::ScheduleWindowEnd => {
-            vec!["schedule_window_end".to_string()]
-        }
-        AutomationTriggerConditionConfig::FocusStarted => vec!["focus_started".to_string()],
-        AutomationTriggerConditionConfig::FocusCompleted => vec!["focus_completed".to_string()],
-        AutomationTriggerConditionConfig::BreakStarted => vec!["break_started".to_string()],
-        AutomationTriggerConditionConfig::BreakCompleted => vec!["break_completed".to_string()],
-        AutomationTriggerConditionConfig::Time { days, at } => {
-            let mut keys = Vec::new();
-            for day in days {
-                if let Some(day_index) = weekday_token_to_index(day) {
-                    keys.push(format!(
-                        "time:{}@{}",
-                        weekday_token_from_index(day_index),
-                        at
-                    ));
-                }
-            }
-            keys.sort();
-            keys.dedup();
-            keys
-        }
-    }
-}
-
-fn format_automation_trigger_conflict_key(key: &str) -> String {
-    if let Some(rest) = key.strip_prefix("time:") {
-        return format!("time trigger `{rest}`");
-    }
-    format!("event trigger `{key}`")
-}
-
-#[derive(Debug, Clone)]
-struct AutomationTriggerConflictRule {
-    index: usize,
-    action: AutomationTriggerActionConfig,
-}
-
-fn automation_trigger_actions_conflict(
-    _existing: &AutomationTriggerActionConfig,
-    _candidate: &AutomationTriggerActionConfig,
-) -> bool {
-    // Automation trigger matrix: overlapping trigger identities always conflict today,
-    // regardless of action type. We keep this explicit helper to evolve policy later.
-    true
-}
-
-fn format_automation_trigger_action(action: &AutomationTriggerActionConfig) -> &'static str {
-    match action {
-        AutomationTriggerActionConfig::StartFocus => "start_focus",
-        AutomationTriggerActionConfig::DelayScheduleStart { .. } => "delay_schedule_start",
-        AutomationTriggerActionConfig::ApplyDefaults { .. } => "apply_defaults",
-    }
-}
-
-fn normalize_weekday_profile_rules(
-    rules: &[WeekdayProfileRuleConfig],
-    blocklist_profiles: &[BlocklistProfileConfig],
-    session_templates: &[SessionTemplateConfig],
-) -> Vec<WeekdayProfileRuleConfig> {
-    let mut normalized_by_day: [Option<WeekdayProfileRuleConfig>; 7] =
-        std::array::from_fn(|_| None);
-    for rule in rules {
-        let Some(normalized) = rule.normalized_with_context(blocklist_profiles, session_templates)
-        else {
-            continue;
-        };
-        let Some(day_index) = weekday_token_to_index(&normalized.day) else {
-            continue;
-        };
-        normalized_by_day[day_index] = Some(normalized);
-    }
-    normalized_by_day.into_iter().flatten().collect()
-}
-
-fn normalize_weekday_token(value: &str) -> Option<String> {
-    weekday_token_to_index(value).map(|index| weekday_token_from_index(index).to_string())
-}
-
-fn normalize_trigger_days(days: &[String]) -> Option<Vec<String>> {
-    let mut normalized = Vec::new();
-    let mut seen = HashSet::new();
-    for day in days {
-        let Some(index) = weekday_token_to_index(day) else {
-            continue;
-        };
-        let token = weekday_token_from_index(index).to_string();
-        if seen.insert(token.clone()) {
-            normalized.push(token);
-        }
-    }
-    if normalized.is_empty() {
-        return None;
-    }
-    normalized.sort_by_key(|day| weekday_token_to_index(day).unwrap_or(usize::MAX));
-    Some(normalized)
-}
-
-fn weekday_token_to_index(value: &str) -> Option<usize> {
-    match value.trim().to_ascii_lowercase().as_str() {
-        "mon" | "monday" => Some(0),
-        "tue" | "tues" | "tuesday" => Some(1),
-        "wed" | "wednesday" => Some(2),
-        "thu" | "thurs" | "thursday" => Some(3),
-        "fri" | "friday" => Some(4),
-        "sat" | "saturday" => Some(5),
-        "sun" | "sunday" => Some(6),
-        _ => None,
-    }
-}
-
-fn weekday_token_from_index(index: usize) -> &'static str {
-    match index {
-        0 => "mon",
-        1 => "tue",
-        2 => "wed",
-        3 => "thu",
-        4 => "fri",
-        5 => "sat",
-        6 => "sun",
-        _ => "mon",
-    }
-}
-
-fn normalize_blocklist_profiles(
-    profiles: &[BlocklistProfileConfig],
-    legacy_blocked_sites: &[String],
-) -> Vec<BlocklistProfileConfig> {
-    let mut normalized = Vec::new();
-    let mut seen_names = HashSet::new();
-
-    for profile in profiles {
-        let base_name =
-            normalize_nonempty_or_default_string(&profile.name, &default_blocklist_profile_name());
-        let name = make_unique_profile_name(&base_name, &mut seen_names);
-        let categories = normalize_blocklist_categories(
-            &profile.categories,
-            &profile.sites,
-            &profile.allowlist_sites,
-        );
-        let selected_category =
-            normalize_selected_blocklist_category(&profile.selected_category, &categories);
-        let (sites, allowlist_sites) = flatten_blocklist_categories(&categories);
-        normalized.push(BlocklistProfileConfig {
-            name,
-            sites,
-            allowlist_sites,
-            categories,
-            selected_category,
-        });
-    }
-
-    if normalized.is_empty() {
-        let categories = normalize_blocklist_categories(&[], legacy_blocked_sites, &[]);
-        let selected_category = normalize_selected_blocklist_category("", categories.as_slice());
-        let (sites, allowlist_sites) = flatten_blocklist_categories(&categories);
-        return vec![BlocklistProfileConfig {
-            name: default_blocklist_profile_name(),
-            sites,
-            allowlist_sites,
-            categories,
-            selected_category,
-        }];
-    }
-
-    normalized
-}
-
-fn normalize_selected_blocklist_profile(
-    selected_name: &str,
-    profiles: &[BlocklistProfileConfig],
-) -> String {
-    let selected_name = selected_name.trim();
-    if selected_name.is_empty() {
-        return profiles
-            .first()
-            .map(|profile| profile.name.clone())
-            .unwrap_or_else(default_blocklist_profile_name);
-    }
-
-    if let Some(profile) = profiles
-        .iter()
-        .find(|profile| profile.name.eq_ignore_ascii_case(selected_name))
-    {
-        profile.name.clone()
-    } else {
-        profiles
-            .first()
-            .map(|profile| profile.name.clone())
-            .unwrap_or_else(default_blocklist_profile_name)
-    }
-}
-
-fn make_unique_profile_name(base_name: &str, seen_names: &mut HashSet<String>) -> String {
-    if seen_names.insert(base_name.to_ascii_lowercase()) {
-        return base_name.to_string();
-    }
-
-    let mut suffix = 2usize;
-    loop {
-        let candidate = format!("{base_name} ({suffix})");
-        if seen_names.insert(candidate.to_ascii_lowercase()) {
-            return candidate;
-        }
-        suffix += 1;
-    }
-}
-
-fn normalize_blocklist_categories(
-    categories: &[BlocklistCategoryConfig],
-    legacy_sites: &[String],
-    legacy_allowlist_sites: &[String],
-) -> Vec<BlocklistCategoryConfig> {
-    let mut normalized = Vec::new();
-    let mut seen_names = HashSet::new();
-    for category in categories {
-        let base_name = normalize_nonempty_or_default_string(
-            &category.name,
-            &default_blocklist_category_name(),
-        );
-        let name = make_unique_profile_name(&base_name, &mut seen_names);
-        normalized.push(BlocklistCategoryConfig {
-            name,
-            sites: category.sites.clone(),
-            allowlist_sites: category.allowlist_sites.clone(),
-        });
-    }
-
-    if normalized.is_empty() {
-        return vec![BlocklistCategoryConfig {
-            name: default_blocklist_category_name(),
-            sites: legacy_sites.to_vec(),
-            allowlist_sites: legacy_allowlist_sites.to_vec(),
-        }];
-    }
-
-    if !legacy_sites.is_empty() || !legacy_allowlist_sites.is_empty() {
-        let target_index = if let Some(index) = normalized.iter().position(|category| {
-            category
-                .name
-                .eq_ignore_ascii_case(&default_blocklist_category_name())
-        }) {
-            index
-        } else {
-            normalized.push(BlocklistCategoryConfig {
-                name: default_blocklist_category_name(),
-                sites: Vec::new(),
-                allowlist_sites: Vec::new(),
-            });
-            normalized.len().saturating_sub(1)
-        };
-        merge_unique_case_insensitive(&mut normalized[target_index].sites, legacy_sites);
-        merge_unique_case_insensitive(
-            &mut normalized[target_index].allowlist_sites,
-            legacy_allowlist_sites,
-        );
-    }
-
-    normalized
-}
-
-fn merge_unique_case_insensitive(target: &mut Vec<String>, source: &[String]) {
-    let mut seen: HashSet<String> = target
-        .iter()
-        .map(|value| value.to_ascii_lowercase())
-        .collect();
-    for value in source {
-        let key = value.to_ascii_lowercase();
-        if seen.insert(key) {
-            target.push(value.clone());
-        }
-    }
-}
-
-fn normalize_selected_blocklist_category(
-    selected_name: &str,
-    categories: &[BlocklistCategoryConfig],
-) -> String {
-    let selected_name = selected_name.trim();
-    if selected_name.is_empty() {
-        return categories
-            .first()
-            .map(|category| category.name.clone())
-            .unwrap_or_else(default_blocklist_category_name);
-    }
-
-    if let Some(category) = categories
-        .iter()
-        .find(|category| category.name.eq_ignore_ascii_case(selected_name))
-    {
-        category.name.clone()
-    } else {
-        categories
-            .first()
-            .map(|category| category.name.clone())
-            .unwrap_or_else(default_blocklist_category_name)
-    }
-}
-
-fn flatten_blocklist_categories(
-    categories: &[BlocklistCategoryConfig],
-) -> (Vec<String>, Vec<String>) {
-    let mut sites = Vec::new();
-    let mut allowlist_sites = Vec::new();
-    let mut seen_sites = HashSet::new();
-    let mut seen_allowlist_sites = HashSet::new();
-
-    for category in categories {
-        for site in &category.sites {
-            let key = site.to_ascii_lowercase();
-            if seen_sites.insert(key) {
-                sites.push(site.clone());
-            }
-        }
-        for site in &category.allowlist_sites {
-            let key = site.to_ascii_lowercase();
-            if seen_allowlist_sites.insert(key) {
-                allowlist_sites.push(site.clone());
-            }
-        }
-    }
-
-    (sites, allowlist_sites)
 }
 
 #[cfg(test)]
